@@ -14,10 +14,12 @@
     },
     "minimum-stability": "stable",
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=8.0",
         "yiisoft/yii2": "~2.0.14",
         "yiisoft/yii2-bootstrap4": "~2.0.0",
-        "yiisoft/yii2-swiftmailer": "~2.0.0 || ~2.1.0"
+        "yiisoft/yii2-swiftmailer": "~2.0.0 || ~2.1.0",
+        "yiisoft/yii2-bootstrap5": "^2.0",
+        "rmrevin/yii2-fontawesome": "^3.7"
     },
     "require-dev": {
         "yiisoft/yii2-debug": "~2.1.0",
